a few JSON refactors, no longer need to define data values
for items that are default at data value 0.  Also, prevents a crash if item is not defined.  One new thing: if you want all items in a tab to override creative, you can add "override_creative": true to the root of the tab and you don't need it on individual items